Removed readonly from user dto. Separated validators from dto's.
