+++ Backend/src/core/dto/dto_validator.ts
@@ -1,4 +1,4 @@
-export default class DTO {
+export default class DTOValidator {
     private errors?: string[];
 
     protected clearErrors() {
